Switch from trigrams to 4-grams
Quality of play is much more enjoyable this way.
Also I fixed an off by one bug in the weighting of ngram moves.  This
was most noticable when there was only one instance of a trigram among
all quadrigrams, and the computer would always plat the first move it
considered since the computed threshold was always zero.
I noticed this on "upk" where it always played "a".